+++ packages/vite-plugin/src/index.ts
@@ -27,8 +27,18 @@ export function ViteCodeInspectorPlugin(options?: Options) {
       if (options?.close) {
         return false;
       }
-      const isDev = command === 'serve';
-      return isDev;
+      // 自定义 dev 环境判断
+      let isDev: boolean;
+      if (typeof options?.dev === 'function') {
+        isDev = options?.dev();
+      } else {
+        isDev = options?.dev;
+      }
+      if (isDev === false) {
+        return false;
+      } else {
+        return !!isDev || command === 'serve';
+      }
     },
     async transform(code, id) {
       // start server and inject client code to entry file
@@ -40,6 +50,10 @@ export function ViteCodeInspectorPlugin(options?: Options) {
       const [_completePath] = id.split('?', 2); // 当前文件的绝对路径
       const filePath = normalizePath(_completePath);
       const params = new URLSearchParams(id);
+      // 仅对符合正则的生效
+      if (options?.match && !options.match.test(filePath)) {
+        return code;
+      }
 
       const jsxParamList = ['isJsx', 'isTsx', 'lang.jsx', 'lang.tsx'];
       const isJsx =