+++ packages/webpack-plugin/src/index.ts
@@ -28,27 +28,34 @@ const applyLoader = (options: LoaderOptions, compiler: any) => {
   const _compiler = compiler?.compiler || compiler;
   const module = _compiler?.options?.module;
   const rules = module?.rules || module?.loaders || [];
-  rules.push({
-    test: /\.(vue|jsx|tsx|js|ts|mjs|mts)$/,
-    exclude: /node_modules/,
-    use: [
-      { 
-        loader: path.resolve(compatibleDirname, `./loader.js`) ,
-        options,
-      }
-    ],
-    ...(options.enforcePre === false ? {} : { enforce: 'pre' })
-  }, {
-    test: /\.(jsx|tsx|js|ts|mjs|mts)$/,
-    exclude: /node_modules/,
-    use: [
-      { 
-        loader: path.resolve(compatibleDirname, `./inject-loader.js`) ,
-        options,
-      }
-    ],
-    enforce: 'post'
-  });
+  rules.push(
+    {
+      test: options?.match ?? /\.(vue|jsx|tsx|js|ts|mjs|mts)$/,
+      exclude: /node_modules/,
+      use: [
+        {
+          loader: path.resolve(compatibleDirname, `./loader.js`),
+          options,
+        },
+      ],
+      ...(options.enforcePre === false ? {} : { enforce: 'pre' }),
+    },
+    {
+      ...(options?.injectTo
+        ? { resource: options?.injectTo }
+        : {
+            test: /\.(jsx|tsx|js|ts|mjs|mts)$/,
+            exclude: /node_modules/,
+          }),
+      use: [
+        {
+          loader: path.resolve(compatibleDirname, `./inject-loader.js`),
+          options,
+        },
+      ],
+      enforce: 'post',
+    }
+  );
 }
 
 interface Options extends CodeOptions {
@@ -70,8 +77,21 @@ class WebpackCodeInspectorPlugin {
       return;
     }
 
+    // 自定义 dev 环境判断
+    let isDev: boolean;
+    if (typeof this.options?.dev === 'function') {
+      isDev = this.options?.dev();
+    } else {
+      isDev = this.options?.dev;
+    }
+
+    if (isDev === false) {
+      return;
+    }
+
     // 仅在开发环境下使用
     if (
+      !isDev &&
       compiler?.options?.mode !== 'development' &&
       process.env.NODE_ENV !== 'development'
     ) {